+++ crates/planus-types/src/ast.rs
@@ -493,6 +493,10 @@ impl IntegerType {
         }
     }
 
+    pub fn is_unsigned(&self) -> bool {
+        matches!(self, Self::U8 | Self::U16 | Self::U32 | Self::U64)
+    }
+
     pub fn flatbuffer_name(&self) -> &'static str {
         match self {
             IntegerType::U8 => "uint8",