+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for `(bit_flags)` enums, generated as a dependency-free flags newtype
+
 ### Fixed
 
 ### Removed